checkpatch: prefer eth_<foo>_addr over memset(,,ETH_ALEN)
Add 2 new checks on memset calls in the file checkpatch.pl as follows:

replace memset by eth_zero_addr if the second argument is
an address of zeros (0x00). eth_zero_addr is a wrapper function
for memset that takes an address array to set as zero. The size
address has to be ETH_ALEN.

replace memset by eth_broadcast_addr if the second argument is
the broadcast address (0xff). eth_broadcast_addr is a wrapper
function for memset that sets the passed array the broadcast
address. The size of the address has to be ETH_ALEN.

In addition, the regex assigned to the variable $stat has been
restricted to handle memset function calls of the form:
[<identifer> =]* memset(<arg1>, <arg2>, <arg3>)

Cc: Julia Lawall <julia.lawall@lip6.fr>
Cc: Andrew Morton <akpm@linux-foundation.org>
Acked-by: Joe Perches <joe@perches.com>
Signed-off-by: Aya Mahfouz <mahfouz.saif.elyazal@gmail.com>
[joe@perches.com: handled all regular expressions]
Portions-by: Joe Perches<joe@perches.com>